+= doesn't like shorts

+++ src/main/scala/com/aethercraft/factionsmap/Renderer.scala
@@ -13,7 +13,7 @@ class Renderer(p: Plugin) extends MapRenderer(true) {
     (1,0),/*(1,1),*/(1,2),
     (2,0),  (2,1),  (2,2)
   )
-  val callCount = new Array[Short](2048)
+  val callCount = new Array[Int](2048)
   def render(map: MapView, canvas: MapCanvas, player: Player) {
     if (callCount(map.getId) % 128 != 0) return
     val shouldLog = callCount(map.getId) % 512 == 0
